requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.15.4
fitz==0.0.1.dev2
Flask==3.0.3
frontend==0.0.3
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
grpcio==1.65.4
grpcio-status==1.62.3
h11==0.14.0
httplib2==0.22.0
idna==3.7
importlib_metadata==8.2.0
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
looseversion==1.3.0
lxml==5.3.0
MarkupSafe==2.1.5
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==2.0.1
packaging==24.1
pandas==2.2.2
pathlib==1.0.1
proto-plus==1.24.0
protobuf==4.25.4
prov==2.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
PyMuPDF==1.24.9
PyMuPDFb==1.24.9
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytils==0.4.1
pytz==2024.1
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
rsa==4.9
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
tools==0.1.9
tqdm==4.66.5
traits==6.3.2
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.6
Werkzeug==3.0.3
zipp==3.20.0
